power: qpnp-smb2: Reverse the sign of CURRENT_NOW
Userspace expects CURRENT_NOW to report positive value
when battery is charging and negative during discharge.

Added note from Roberto:
On 4.4, '/sys/class/power_supply/battery/current_now' indeed reports a
positive value when battery is charging, while 4.14 before this commit was
reporting a negative value.

The 'dashd' userspace daemon reads the 'current_now' node so it's better
to align to 4.4 on this.

Signed-off-by: Kavya Nunna <knunna@codeaurora.org>